Aelin price refactor
